requirements.txt
SpeechRecognition
pipwin
pyaudio
gTTS
google-generativeai
python-dotenv
streamlit
pyaudio
wave
langchain==0.0.225
ctransformers==0.2.5
sentence-transformers==2.2.2
pypdf
flask
faiss-cpu
unstructured
langchain_community
langchain-text-splitters
accelerate==0.20.3
selenium
playwright
ffmpeg-downloader
pydub


-e .